requirements.txt
streamlit
pytesseract
pillow
pdfplumber
python-docx
